requirements.txt
discord.py
discord.py[voice]
youtube_dl>=2020.03.24
pymongo[srv]
PyNaCl